Control.Failure

Description

Type classes for returning failures.

Note: This module used to contain a lot more functionality, but I believe it was unused functionality. If you want any of it back, just email me.

Type class

class Monad f => Failure e f where

Methods

failure :: e -> f v

General exceptions

exception :: (Exception e, Failure SomeException m) => e -> m a

Convert to a SomeException via toException before calling failure.