Failure.Error

Description

Concrete Error types

Since: failure-0.1.0

Documentation

data Error a

The concrete error type

Since: failure-0.1.0

code :: forall a. Lens' (Error a) a

description :: forall a. Lens' (Error a) String

err

Arguments

:: a
The error code

-> String
The description

-> Error a
The resulting Error

This is a convenient way to turn a string into an error value that can be passed around, if you do not want to create a new Fail type for this use case.

Since: failure-0.1.0

nextCause :: forall a. Lens' (Error a) (Maybe (Error a))