Failure.Fail

Description

The interface for Fail types

Since: failure-0.1.0

Documentation

class Fail a where

The main class of Fails.

Types which implement Fail are called failures.

backtrace evaluates to the ordered list of failures

cause returns the underlying cause of this failure, if it is an error that wraps other errors. Returns Nothing if this failure does not have another error as its underlying cause.

Since: failure-0.1.0

Methods

backtrace :: a -> [a]

cause :: a -> Maybe a