Failure.Fail

Description

The interface for Fail types

Since: failure-0.1.0

Documentation

class Fail a where

The main class of Fails.

Types which implement Fail are called failures.

backtrace evaluates to the ordered list of failures

cause returns the underlying cause of this failure, if it is an error that wraps other errors. Returns Nothing if this failure does not have another error as its underlying cause.

+> adds an Error to the chain. This sets the seconds error cause to the first one and prepends it therefore to the overall error-chain.

Since: failure-0.1.0

Methods

backtrace

Arguments

:: a
The input failure

-> [a]
The resulting list of failures as backtrace

cause

Arguments

:: a
The input failure

-> Maybe a
The cause failure

(+>)

Arguments

:: a
The reason for the resulting failure

-> a
The failure to append

-> a
The resulting failure